Sync with upstream script to properly handle symlinks (#27)
I ran a comparison of the chromium's sysroots with ours and found out,
that symlinks look differently, e.g.
- chromium: `libutil.so -> ../../..//lib/x86_64-linux-gnu/libutil.so.1`
- ours (before this change): `libutil.so ->
/lib/x86_64-linux-gnu/libutil.so.1`

This might have worked in the past with Bazel 7, but with Bazel 8 in
some situations (maybe due to Bazel 8's stricter sandboxing) it just
fails. I was not able to reproduce it locally, it only happens in CI
with enabled Bazel Cache.

This change enables the relative symlink handling instead of using
absolute paths and syncs with the upstream script of the chromium
project.

(`lapack` and `blas` seem to be in the correct shape: `liblapack.a ->
./lapack/liblapack.a`)

Tested in https://github.com/swift-nav/rules_swiftnav/pull/200 (see
testplan in PR description)